Avoid reading more than one message at a time
Instead of assuming there is an infinite amount of messages in the
stream, read just one at a time into the buffer supplied. If the buffer
is not big enough, remember how much needs to be read. If the buffer is
bigger than the message, read ONLY that message into the buffer and
await another call to Read.